Moved growth settings out of the settings quarantine
no ref

The growth settings components now live beside their acceptance tests in
settings/growth, the third area to leave settings/app. Moving them
enables the type-aware lint rules the rest of admin uses, so the same
debt is cleared for these files: async handlers passed to void callbacks
are wrapped, fire-and-forget promises are marked, and the non-component
exports move into their own modules — the search keywords, the offer
cadence/duration/discount helpers (into the existing offer-helpers) and
the recommendation description validation — so the component files only
export components.
