+++ apps/admin/src/settings/app/components/settings.tsx
@@ -4,7 +4,7 @@ import AdvancedSettings from './settings/advanced/advanced-settings';
 import EmailSettings from './settings/email/email-settings';
 import Emails from './settings/email/emails';
 import GeneralSettings from '@/settings/general/general-settings';
-import GrowthSettings from './settings/growth/growth-settings';
+import GrowthSettings from '@/settings/growth/growth-settings';
 import MembershipSettings from '@/settings/membership/membership-settings';
 import SiteSettings from './settings/site/site-settings';
 import useFeatureFlag from '@/settings/app/hooks/use-feature-flag';
