+++ apps/admin/src/settings/growth/embed-signup/embed-signup-sidebar.tsx
