+++ apps/admin/src/settings/growth/growth-settings.tsx
@@ -6,15 +6,8 @@ import React from 'react';
 import Recommendations from './recommendations';
 import SearchableSection from '@/settings/app/components/searchable-section';
 import {checkStripeEnabled} from '@tryghost/admin-x-framework/api/settings';
+import {searchKeywords} from './search-keywords';
 import {useGlobalData} from '@/settings/app/components/providers/global-data-provider';
-
-export const searchKeywords = {
-    network: ['growth', 'network', 'activitypub', 'blog', 'fediverse', 'sharing'],
-    explore: ['ghost explore', 'explore', 'growth', 'share', 'list', 'listing'],
-    recommendations: ['growth', 'recommendations', 'recommend', 'blogroll'],
-    embedSignupForm: ['growth', 'embeddable signup form', 'embeddable form', 'embeddable sign up form', 'embeddable sign up'],
-    offers: ['growth', 'offers', 'discounts', 'coupons', 'promotions']
-};
 
 const GrowthSettings: React.FC = () => {
     const {config, settings} = useGlobalData();
