+++ apps/admin/src/settings/growth/offers/offer-helpers.ts
@@ -1,3 +1,6 @@
+import {type Tier} from '@tryghost/admin-x-framework/api/tiers';
+import {currencyToDecimal, getSymbol} from '@/settings/app/utils/currency';
+import {formatNumber} from '@tryghost/shade/utils';
 import {getGhostPaths} from '@tryghost/admin-x-framework/helpers';
 
 const MAX_RETENTION_OFFER_NAME_LENGTH = 40;
@@ -63,4 +66,54 @@ export const generateRetentionOfferName = (input: {amount: number; duration: str
 
 export {
     MAX_RETENTION_OFFER_NAME_LENGTH
+};
+
+export type OfferType = 'percent' | 'fixed' | 'trial';
+
+export const getOfferCadence = (cadence: string): string => {
+    return cadence === 'month' ? 'monthly' : 'yearly';
+};
+
+export const getOfferDuration = (duration: string): string => {
+    return (duration === 'once' ? 'First payment' : duration === 'repeating' ? 'Repeating' : 'Forever');
+};
+
+export const getOfferDiscount = (type: string, amount: number, cadence: string, currency: string, tier: Tier | undefined): {discountOffer: string, originalPriceWithCurrency: string, updatedPriceWithCurrency: string} => {
+    let discountOffer = '';
+    const originalPrice = cadence === 'month' ? tier?.monthly_price ?? 0 : tier?.yearly_price ?? 0;
+    let updatedPrice = originalPrice;
+
+    const formatToTwoDecimals = (num: number): number => parseFloat(num.toFixed(2));
+    const formatPrice = (num: number): string => formatNumber(formatToTwoDecimals(currencyToDecimal(num)), {maximumFractionDigits: 2});
+
+    let originalPriceWithCurrency = getSymbol(currency) + formatPrice(originalPrice);
+
+    switch (type) {
+    case 'percent':
+        discountOffer = `${formatNumber(amount)}% off`;
+        updatedPrice = originalPrice - ((originalPrice * amount) / 100);
+        break;
+    case 'fixed':
+        discountOffer = `${formatPrice(amount)} ${currency} off`;
+        updatedPrice = originalPrice - amount;
+        break;
+    case 'trial':
+        discountOffer = `${formatNumber(amount)} days free`;
+        originalPriceWithCurrency = '';
+        break;
+    default:
+        break;
+    };
+
+    if (updatedPrice < 0) {
+        updatedPrice = 0;
+    }
+
+    const updatedPriceWithCurrency = getSymbol(currency) + formatPrice(updatedPrice);
+
+    return {
+        discountOffer,
+        originalPriceWithCurrency,
+        updatedPriceWithCurrency
+    };
 };
