+++ apps/admin/src/settings/growth/recommendations/add-recommendation-modal-confirm.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import RecommendationDescriptionForm, {validateDescriptionForm} from './recommendation-description-form';
+import RecommendationDescriptionForm from './recommendation-description-form';
+import {validateDescriptionForm} from './recommendation-validation';
 import trackEvent from '@/settings/app/utils/analytics';
 import {Button} from '@tryghost/shade/components';
 import {type EditOrAddRecommendation, useAddRecommendation} from '@tryghost/admin-x-framework/api/recommendations';
