+++ apps/admin/src/settings/growth/recommendations/recommendation-description-form.tsx
@@ -5,6 +5,7 @@ import {type ErrorMessages} from '@tryghost/admin-x-framework/hooks';
 import {Field, FieldDescription, FieldError, FieldGroup, FieldLabel, Input, Textarea} from '@tryghost/shade/components';
 import {Text} from '@tryghost/shade/primitives';
 import {formatNumber} from '@tryghost/shade/utils';
+import {validateDescriptionForm} from './recommendation-validation';
 
 interface Props<T extends EditOrAddRecommendation> {
     showURL?: boolean,
@@ -14,39 +15,6 @@ interface Props<T extends EditOrAddRecommendation> {
     clearError?: (key: keyof ErrorMessages) => void,
     setErrors: (errors: ErrorMessages) => void
 }
-
-export const validateDescriptionFormField = function (errors: ErrorMessages, field: 'title'|'description', value: string|null) {
-    const cloned = {...errors};
-    switch (field) {
-    case 'title':
-        if (!value) {
-            cloned.title = 'Title is required';
-        } else {
-            delete cloned.title;
-        }
-        break;
-    case 'description':
-        if (value && value.length > 200) {
-            cloned.description = 'Description cannot be longer than 200 characters';
-        } else {
-            delete cloned.description;
-        }
-        break;
-    default: {
-        // Will throw a compile error if we forget to add a case for a field
-        const f: never = field;
-        throw new Error(`Unknown field ${f}`);
-    }
-    }
-    return cloned;
-};
-
-export const validateDescriptionForm = function (formState: EditOrAddRecommendation) {
-    let newErrors: ErrorMessages = {};
-    newErrors = validateDescriptionFormField(newErrors, 'title', formState.title);
-    newErrors = validateDescriptionFormField(newErrors, 'description', formState.description);
-    return newErrors;
-};
 
 function RecommendationDescriptionForm<T extends EditOrAddRecommendation>({showURL, formState, updateForm, errors, clearError, setErrors}: Props<T>) {
     const [descriptionLength, setDescriptionLength] = React.useState(formState?.description?.length || 0);
