+++ apps/admin/src/settings/growth/recommendations/recommendation-icon.tsx
