+++ apps/admin/src/settings/growth/search-keywords.ts
@@ -1,0 +1,7 @@
+export const searchKeywords = {
+    network: ['growth', 'network', 'activitypub', 'blog', 'fediverse', 'sharing'],
+    explore: ['ghost explore', 'explore', 'growth', 'share', 'list', 'listing'],
+    recommendations: ['growth', 'recommendations', 'recommend', 'blogroll'],
+    embedSignupForm: ['growth', 'embeddable signup form', 'embeddable form', 'embeddable sign up form', 'embeddable sign up'],
+    offers: ['growth', 'offers', 'discounts', 'coupons', 'promotions']
+};
